% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.gamma.R
\name{AD.gamma}
\alias{AD.gamma}
\title{EDF statistics A^2 for Gamma Distribution}
\usage{
AD.gamma(x, parameter = estimate.gamma(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameters of Gamma distribution}
}
\value{
AD.gamma gives Anderson-Darling statistic of a uniform sample.
}
\description{
Compute Anderson-Darling statistic A^2 for an iid sample, x, to test for the Gamma distribution with parameters unknown.
Estimate parameters by ML using "estimate.gamma" by default.
}
\examples{
x = rgamma(100,1,2)
AD.gamma(x)
AD.gamma(x,c(1,1/2))
}
