% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.logistic.covmat.R
\name{AD.logistic.covmat}
\alias{AD.logistic.covmat}
\title{AD.logistic.covmat}
\usage{
AD.logistic.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
AD.logistic.covmat
}
