% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.logistic.pvalue.R
\name{AD.logistic.pvalue}
\alias{AD.logistic.pvalue}
\title{P-value of EDF statistics A^2 for Logistic Distribution}
\usage{
AD.logistic.pvalue(a, neig = 100, verbose = FALSE)
}
\arguments{
\item{a}{A^2 for Logistic Distribution}

\item{neig}{number of eigenvalues}

\item{verbose}{logical; if TRUE, print warning messages}
}
\value{
AD.logistic.pvalue gives p-value of the Anderson-Darling statistic of a uniform sample.
}
\description{
Compute p-value of the given Anderson-Darling statistic A^2
}
\examples{
x = rlogis(1000)
asq = AD.logistic(x)
AD.logistic.pvalue(asq)
}
