% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.logistic.R
\name{AD.logistic}
\alias{AD.logistic}
\title{Anderson-Darling Statistics for the Logistic Distribution}
\usage{
AD.logistic(x, parameter = estimate.logistic(x))
}
\arguments{
\item{x}{numerical vector}

\item{parameter}{}
}
\value{
float
}
\description{
Computes Anderson-Darling statistic A^2 for an iid sample, x, to test for the Logistic distribution with parameters unknown.
Estimate parameters by ML using "estimate.logistic"
}
\examples{

}
