% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.normal.pvalue.R
\name{AD.normal.pvalue}
\alias{AD.normal.pvalue}
\title{AD.normal.pvalue}
\usage{
AD.normal.pvalue(a, nd = 9, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
AD.normal.pvalue
}
\examples{

}
