% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.normal.R
\name{AD.normal}
\alias{AD.normal}
\title{EDF statistics A^2 for Normal Distribution}
\usage{
AD.normal(x, parameter = estimate.normal(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameters of Normal distribution}
}
\value{
AD.normal gives Anderson-Darling statistic of a uniform sample.
}
\description{
Compute Anderson-Darling statistic A^2 for an iid sample, x, to test for the Normal distribution with parameters unknown.
Estimate parameters by ML using "estimate.normal" by default.
}
\examples{
x= rnorm(1000)
AD.normal(x)
AD.normal(x,c(0,1))
}
