% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.R
\name{AD}
\alias{AD}
\title{EDF statistics A^2}
\usage{
AD(z)
}
\arguments{
\item{z}{vector of numbers supposed to be 0 and 1}
}
\value{
AD gives Anderson-Darling statistic of a uniform sample.
}
\description{
Given probability integral transforms, compute the Anderson-Darling goodness-of-fit statistic
for testing uniformity on the unit interval.
}
