% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.weibull.pvalue.R
\name{AD.weibull.pvalue}
\alias{AD.weibull.pvalue}
\title{AD.weibull.pvalue}
\usage{
AD.weibull.pvalue(a, neig = 100, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
AD.weibull.pvalue
}
