% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.weibull.R
\name{AD.weibull}
\alias{AD.weibull}
\title{Anderson-Darling Statistics for the Weibull Distribution}
\usage{
AD.weibull(x, parameter = estimate.weibull(x))
}
\arguments{
\item{x}{numerical vector}

\item{parameter}{}
}
\value{
float
}
\description{
Computes Anderson-Darling statistic A^2 for an iid sample, x, to test for the Weibull distribution with parameters unknown.
Estimate parameters by ML using "estimate.weibull"
}
\examples{

}
