% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.gamma.regression.R
\name{cdf.gamma.regression}
\alias{cdf.gamma.regression}
\title{cdf.gamma.regression}
\usage{
cdf.gamma.regression(y, x, theta.hat)
}
\arguments{
\item{theta.hat}{}
}
\value{

}
\description{
cdf.gamma.regression
}
