% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.laplace.regression.R
\name{cdf.laplace.regression}
\alias{cdf.laplace.regression}
\title{cdf.laplace.regression}
\usage{
cdf.laplace.regression(fit, data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
cdf.laplace.regression
}
