% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.normal.regression.R
\name{cdf.normal.regression}
\alias{cdf.normal.regression}
\title{cdf.normal.regression}
\usage{
cdf.normal.regression(y, x, theta.hat)
}
\arguments{
\item{theta.hat}{}
}
\value{

}
\description{
cdf.normal.regression
}
