% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.gamma.covmat.R
\name{CvM.gamma.covmat}
\alias{CvM.gamma.covmat}
\title{CvM.gamma.covmat}
\usage{
CvM.gamma.covmat(n, shape)
}
\arguments{
\item{n}{number of eigenvalues}

\item{shape}{shape parameter of Gamma distribution}
}
\value{

}
\description{
CvM.gamma.covmat
}
