% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.gamma.eigen.R
\name{CvM.gamma.eigen}
\alias{CvM.gamma.eigen}
\title{CvM.gamma.eigen}
\usage{
CvM.gamma.eigen(n, shape)
}
\arguments{
\item{n}{number of eigenvalues}

\item{shape}{shape parameter of Gamma distribution}
}
\value{

}
\description{
CvM.gamma.eigen
}
