% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.gamma.pvalue.R
\name{CvM.gamma.pvalue}
\alias{CvM.gamma.pvalue}
\title{P-value of EDF statistics W^2 for Gamma Distribution}
\usage{
CvM.gamma.pvalue(w, shape, neig = 100, verbose = FALSE)
}
\arguments{
\item{w}{W^2 for Gamma Distribution}

\item{shape}{shape parameter of Gamma distribution}

\item{neig}{number of eigenvalues}

\item{verbose}{logical; if TRUE, print warning messages}
}
\value{
CvM.gamma.pvalue gives p-value of the Cramer-von Mises statistic of a uniform sample.
}
\description{
Compute p-value of the given Cramer-von Mises statistic W^2
}
\examples{
x = rgamma(100,1,1)
wsq = CvM.gamma(x)
CvM.gamma.pvalue(wsq,1)
}
