% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.gamma.R
\name{CvM.gamma}
\alias{CvM.gamma}
\title{Cramer Von Mises Statistics for the Gamma Distribution}
\usage{
CvM.gamma(x, parameter = estimate.gamma(x))
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
Computes Cramer Von Mises statistic W^2 for an iid sample, x, to test for the Gamma distribution with parameters unknown.
Estimate parameters by default ML using "estimate.gamma"
}
\examples{

}
