% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.laplace.covmat.R
\name{CvM.laplace.covmat}
\alias{CvM.laplace.covmat}
\title{CvM.laplace.covmat}
\usage{
CvM.laplace.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
CvM.laplace.covmat
}
