% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.logistic.eigen.R
\name{CvM.logistic.eigen}
\alias{CvM.logistic.eigen}
\title{CvM.logistic.eigen}
\usage{
CvM.logistic.eigen(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
CvM.logistic.eigen
}
