% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.logistic.pvalue.R
\name{CvM.logistic.pvalue}
\alias{CvM.logistic.pvalue}
\title{CvM.logistic.pvalue}
\usage{
CvM.logistic.pvalue(w, nd = 9, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
CvM.logistic.pvalue
}
