% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.logistic.R
\name{CvM.logistic}
\alias{CvM.logistic}
\title{Cramer Von Mises Statistics for the Logistic Distribution}
\usage{
CvM.logistic(x, parameter = estimate.logistic(x))
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
Computes Cramer Von Mises statistic W^2 for an iid sample, x, to test for the Logistic distribution with parameters unknown.
Estimate parameters by default ML using "estimate.logistic"
}
\examples{

}
