% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.normal.pvalue.R
\name{CvM.normal.pvalue}
\alias{CvM.normal.pvalue}
\title{CvM.normal.pvalue}
\usage{
CvM.normal.pvalue(w, nd = 7, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
CvM.normal.pvalue
}
\examples{

}
