% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.poisson.R
\name{CvM.poisson}
\alias{CvM.poisson}
\title{Cramer Von Mises Statistics for the Poisson Distribution}
\usage{
CvM.poisson(x, eps = 10^(-9))
}
\arguments{
\item{eps}{how much prob you are willing to omit}
}
\value{

}
\description{
Cramer Von Mises Statistics for the Poisson Distribution
}
