% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.R
\name{CvM}
\alias{CvM}
\title{EDF statistics W^2}
\usage{
CvM(z)
}
\arguments{
\item{z}{vector of numbers supposed to be 0 and 1}
}
\value{
CvM gives Cramer-von Mises statistic of a uniform sample.
}
\description{
Given probability integral transforms, compute the Cramer-von Mises goodness-of-fit statistic
for testing uniformity on the unit interval. It is a component of doing other tests.
}
