% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.weibull.pvalue.R
\name{CvM.weibull.pvalue}
\alias{CvM.weibull.pvalue}
\title{CvM.weibull.pvalue}
\usage{
CvM.weibull.pvalue(w, neig = 100, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
CvM.weibull.pvalue
}
