% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.weibull.R
\name{CvM.weibull}
\alias{CvM.weibull}
\title{Cramer Von Mises Statistics for the Weibull Distribution}
\usage{
CvM.weibull(x, parameter = estimate.weibull(x))
}
\arguments{
\item{parameter}{}
}
\value{

}
\description{
Computes Cramer Von Mises statistic W^2 for an iid sample, x, to test for the Weibull distribution with parameters unknown.
Estimate parameters by default ML using "estimate.weibull"
}
\examples{

}
