% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.exp.R
\name{estimate.exp}
\alias{estimate.exp}
\title{MLE for Exponential Distribution}
\usage{
estimate.exp(x)
}
\arguments{
\item{x}{random sample from the Exponential distirbution}
}
\value{

}
\description{
Computes the mle of the scale parameter for a sample x from the Exponential distribution.
}
\examples{

}
