% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.R
\name{estimate.gamma}
\alias{estimate.gamma}
\title{MLE for Gamma Distribution}
\usage{
estimate.gamma(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Computes the mle of the shape, alpha, and scale, beta, for a sample x from the Gamma distribution.
}
\examples{

}
