% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.regression.R
\name{estimate.gamma.regression}
\alias{estimate.gamma.regression}
\title{estimate.gamma.regression}
\usage{
estimate.gamma.regression(x, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
estimate.gamma.regression
}
