% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.laplace.R
\name{estimate.laplace}
\alias{estimate.laplace}
\title{MLE for Laplace (double exponential) Distribution}
\usage{
estimate.laplace(x)
}
\arguments{
\item{x}{random sample}
}
\value{
estimated location and scale parameters of the Laplace distribution.
}
\description{
Estimate location and scale parameters of the Laplace distribution by the method of maximum likelihood.
}
\examples{
x=runif(10,-5,5)
estimate.laplace(x)
}
