% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.logistic.R
\name{estimate.logistic}
\alias{estimate.logistic}
\title{MLE for Logistic Distribution}
\usage{
estimate.logistic(x, eps = 1e-07, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
MLE for Logistic Distribution
}
\examples{

}
