% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.normal.R
\name{estimate.normal}
\alias{estimate.normal}
\title{MLE for Normal Distribution}
\usage{
estimate.normal(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Computes the mean and variance of a sample x from the normal distribution.
}
