% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.normal.regression.R
\name{estimate.normal.regression}
\alias{estimate.normal.regression}
\title{estimate.normal.regression}
\usage{
estimate.normal.regression(x, y, fit.intercept = TRUE)
}
\arguments{
\item{fit.intercept}{}
}
\value{

}
\description{
estimate.normal.regression
}
