% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.weibull.R
\name{estimate.weibull}
\alias{estimate.weibull}
\title{MLE for Weibull Distribution}
\usage{
estimate.weibull(x)
}
\arguments{
\item{x}{numerical vector}
}
\value{
vector of two parameters, the first alpha, the second beta
}
\description{
MLE for Weibull Distribution
}
\examples{

}
