% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.gamma.bootstrap.R
\name{gof.gamma.bootstrap}
\alias{gof.gamma.bootstrap}
\title{EDF Goodness-of-Fit tests for Gamma Distribution by Bootstrap}
\usage{
gof.gamma.bootstrap(x, M = 10000)
}
\arguments{
\item{x}{random sample}

\item{M}{number of bootstrap}
}
\value{
gof.gamma.bootstrap computes Cramer-von Mises and Anderson-Darling statistics and their P-values by bootstrap.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Gamma
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics. P-values are calculated by M bootstrap.
}
\examples{
x=rgamma(10,1)
gof.gamma.bootstrap(x,M=1000)
gof.gamma(x)
}
