% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.laplace.bootstrap.R
\name{gof.laplace.bootstrap}
\alias{gof.laplace.bootstrap}
\title{EDF Goodness-of-Fit tests for Laplace Distribution by Bootstrap}
\usage{
gof.laplace.bootstrap(x, M = 10000)
}
\arguments{
\item{x}{random sample}

\item{M}{number of bootstrap}
}
\value{
gof.laplace.bootstrap computes Cramer-von Mises and Anderson-Darling statistics and their P-values by bootstrap.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Laplace
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics. P-values are calculated by M bootstrap.
}
\examples{
library(L1pack)
x= rlaplace(1000,0,1)
gof.laplace.bootstrap(x,M=1000)
gof.laplace(x)
}
