% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.laplace.R
\name{gof.laplace}
\alias{gof.laplace}
\title{EDF Goodness-of-Fit tests for Laplace Distribution}
\usage{
gof.laplace(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{print}{Logical, if TRUE print statistics and P-values}

\item{verbose}{Logical, if TRUE print every step}
}
\value{
gof.laplace computes Cramer-von Mises and Anderson-Darling statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Laplace
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics and their P-values
}
\examples{
library(L1pack)
x= rlaplace(1000,0,1)
gof.laplace(x)
gof.laplace(x,print=TRUE,verbose=TRUE)
}
