% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.normal.bootstrap.R
\name{gof.normal.bootstrap}
\alias{gof.normal.bootstrap}
\title{EDF Goodness-of-Fit tests for Normal Distribution by Bootstrap}
\usage{
gof.normal.bootstrap(x, M = 10000)
}
\arguments{
\item{x}{random sample}

\item{M}{number of bootstrap}
}
\value{
gof.normal.bootstrap computes Cramer-von Mises and Anderson-Darling statistics and their P-values by bootstrap.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics. P-values are calculated by M bootstrap.
}
\examples{
x=rnorm(1000)
gof.normal.bootstrap(x,M=1000)
gof.normal(x)
}
