% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.normal.R
\name{gof.normal}
\alias{gof.normal}
\title{EDF Goodness-of-Fit tests for Normal Distribution}
\usage{
gof.normal(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{print}{Logical, if TRUE print statistics and P-values}

\item{verbose}{Logical, if TRUE print every step}
}
\value{
gof.normal computes Cramer-von Mises and Anderson-Darling statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics and their P-values
}
\examples{
x = rnorm(100,1)
gof.normal(x)
gof.normal(x,print=TRUE,verbose=TRUE)
}
