% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.sandwich.R
\name{gof.sandwich}
\alias{gof.sandwich}
\title{gof.sandwich}
\usage{
gof.sandwich(y, x = NULL, Fdist, thetahat, Score, m = max(n, 100), ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
gof.sandwich
}
\examples{

}
