% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.statistics.only.R
\name{gof.statistics.only}
\alias{gof.statistics.only}
\title{gof.statistics.only}
\usage{
gof.statistics.only(pit, a2 = T, w2 = T, u2 = T)
}
\arguments{
\item{u2}{}
}
\value{

}
\description{
gof.statistics.only
}
