% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.uniform.bootstrap.R
\name{gof.uniform.bootstrap}
\alias{gof.uniform.bootstrap}
\title{gof.uniform.bootstrap}
\usage{
gof.uniform.bootstrap(x, stat = y, m = 10000)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
gof.uniform.bootstrap
}
