% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.weibull.bootstrap.R
\name{gof.weibull.bootstrap}
\alias{gof.weibull.bootstrap}
\title{gof.weibull.bootstrap}
\usage{
gof.weibull.bootstrap(x, M = 10000)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
gof.weibull.bootstrap
}
