% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.gamma.regression.R
\name{score.gamma.regression}
\alias{score.gamma.regression}
\title{score.gamma.regression}
\usage{
score.gamma.regression(y, x, theta.hat)
}
\arguments{
\item{theta.hat}{}
}
\value{

}
\description{
score.gamma.regression
}
\examples{
set.seed(5)
n = 200  #(sample size)
p = 3
beta = c(3,4,5)
alpha = 3
x = rnorm(n*(p-1))
x = c(rep(1,n),x)
x = matrix(x,n,p)
scale=exp(x \%*\% beta) / alpha
y =rgamma(n,shape=alpha,scale=scale)
theta.hat = estimate.gamma.regression(x,y)

}
