% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.weibull.R
\name{score.weibull}
\alias{score.weibull}
\title{Score Function for Weibull Distribution}
\usage{
score.weibull(x, theta)
}
\arguments{
\item{x}{random sample}

\item{theta}{parameter of Weibull distribution}
}
\value{
score.weibull returns an n by 2 matrix whose ith row refers to the ith data point.
}
\description{
Compute the score function for the two parameter Weibull distribution
}
