% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.logistic.R
\name{Watson.logistic}
\alias{Watson.logistic}
\title{EDF statistics U^2 for Logistic Distribution}
\usage{
Watson.logistic(x, parameter = estimate.logistic(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Logistic distribution}
}
\value{
Watson.logistic gives Watson statistic of a uniform sample
}
\description{
Compute Watson statistic U^2 for an iid sample, x, to test for the Logistic distribution with parameters unknown.
Estimate parameters by ML using "estimate.logistic" by default.
}
\examples{
x= rlogis(1000)
estimate.logistic(x)
Watson.logistic(x)
}
