% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.R
\name{Watson}
\alias{Watson}
\title{EDF statistics U^2}
\usage{
Watson(z)
}
\arguments{
\item{z}{vector of numbers supposed to be 0 and 1.}
}
\value{
Watson gives Watson statistic of a uniform sample.
}
\description{
Given probability integral transforms, compute the Watson goodness-of-fit statistic
for testing uniformity on the unit interval. It is also a component of doing other tests.
}
