% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.weibull.R
\name{Watson.weibull}
\alias{Watson.weibull}
\title{EDF statistics U^2 for Weibull Distribution}
\usage{
Watson.weibull(x, parameter = estimate.weibull(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Weibull distribution}
}
\value{
Watson.weibull gives Watson statistic of a uniform sample
}
\description{
Compute Watson statistic U^2 for an iid sample, x, to test for the Weibull distribution with parameters unknown.
Estimate parameters by ML using "estimate.weibull" by default.
}
\examples{
x= rweibull(1000,1)
estimate.weibull(x)
Watson.weibull(x)
}
