% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.gamma.R
\name{AD.gamma}
\alias{AD.gamma}
\title{Anderson-Darling Statistics for the Gamma Distribution}
\usage{
AD.gamma(x, parameter = estimate.gamma(x))
}
\arguments{
\item{x}{numerical vector}

\item{parameter}{}
}
\value{
float
}
\description{
Computes Anderson-Darling statistic A^2 for an iid sample, x, to test for the Gamma distribution with parameters unknown.
Estimate parameters by ML using "estimate.gamma"
}
\examples{

}
