% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.logistic.pvalue.R
\name{AD.logistic.pvalue}
\alias{AD.logistic.pvalue}
\title{AD.logistic.pvalue}
\usage{
AD.logistic.pvalue(a, nd = 9, verbose = F)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
AD.logistic.pvalue
}
