% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.R
\name{AD}
\alias{AD}
\title{EDF statistics A^2}
\usage{
AD(z)
}
\arguments{
\item{z}{Assumes z is a vector with 0 < z[i] < 1 for each i}
}
\value{
float
}
\description{
Computes the Anderson-Darling goodness-of-fit statistic for testing uniformity on the unit interval.
}
\examples{

}
