% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.gamma.pvalue.R
\name{CvM.gamma.pvalue}
\alias{CvM.gamma.pvalue}
\title{CvM.gamma.pvalue}
\usage{
CvM.gamma.pvalue(w, shape, neig = 100, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
CvM.gamma.pvalue
}
