% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.laplace.pvalue.R
\name{CvM.laplace.pvalue}
\alias{CvM.laplace.pvalue}
\title{CvM.laplace.pvalue}
\usage{
CvM.laplace.pvalue(w, neig = 100, verbose = F)
}
\arguments{
\item{verbose}{}
}
\description{
CvM.laplace.pvalue
}
