% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.poisson.pvalue.bootstrap.R
\name{CvM.poisson.pvalue.bootstrap}
\alias{CvM.poisson.pvalue.bootstrap}
\title{CvM.poisson.pvalue.bootstrap}
\usage{
CvM.poisson.pvalue.bootstrap(x, nmc = 200, return.samples = F, print = TRUE)
}
\arguments{
\item{x}{random sample}

\item{print}{}
}
\value{

}
\description{
CvM.poisson.pvalue.bootstrap
}
