% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.R
\name{CvM}
\alias{CvM}
\title{EDF statistics W^2}
\usage{
CvM(z)
}
\arguments{
\item{z}{Assumes z is a vector with 0 < z[i] < 1 for each i}
}
\value{
float
}
\description{
Computes the Cramer-von Mises goodness-of-fit statistic for testing uniformity on the unit interval.
}
\examples{

}
