% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.exp.R
\name{estimate.exp}
\alias{estimate.exp}
\title{MLE for Exponential Distribution}
\usage{
estimate.exp(x)
}
\arguments{
\item{x}{random sample}
}
\value{
estimated scale parameter of the Exponential distribution.
}
\description{
Estimate scale parameter of the Exponential distribution by the method of maximum likelihood.
}
\examples{
x=rexp(10)
estimate.exp(x)
}
