% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.R
\name{estimate.gamma}
\alias{estimate.gamma}
\title{MLE for Gamma Distribution}
\usage{
estimate.gamma(x)
}
\arguments{
\item{x}{random sample}
}
\value{
estimated shape and scale parameters of the Gamma distribution.
}
\description{
Estimate shape and scale parameters of the Gamma distribution by the method of maximum likelihood.
Use the digamma and trigamma functions in Base R and do Newton-Raphson on the profile log-likelihood for alpha
}
\examples{
x=rgamma(10,2,1)
estimate.gamma(x)
}
