% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.laplace.regression.R
\name{estimate.laplace.regression}
\alias{estimate.laplace.regression}
\title{estimate.laplace.regression}
\usage{
estimate.laplace.regression(x, y, fit.intercept = TRUE)
}
\arguments{
\item{fit.intercept}{}
}
\value{

}
\description{
estimate.laplace.regression
}
