% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.logistic.R
\name{estimate.logistic}
\alias{estimate.logistic}
\title{MLE for Logistic Distribution}
\usage{
estimate.logistic(x, eps = 1e-07, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{eps}{stopping criterion}

\item{verbose}{logical; if TRUE, print the warning message}
}
\value{
estimated location and scale parameters of the Logistic distribution.
}
\description{
Estimate location and scale parameters of the Logistic distribution by the method of maximum likelihood.
}
\examples{
x=rlogis(10)
estimate.logistic(x)
}
