% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.weibull.R
\name{estimate.weibull}
\alias{estimate.weibull}
\title{MLE for Weibull Distribution}
\usage{
estimate.weibull(x)
}
\arguments{
\item{x}{random sample}
}
\value{
estimated shape and scale parameters of the Weibull distribution.
}
\description{
Estimate shape and scale parameters of the Weibull distribution by the method of maximum likelihood.
}
\examples{
x=rweibull(10,1)
estimate.weibull(x)
}
