% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.information.normal.R
\name{fisher.information.normal}
\alias{fisher.information.normal}
\title{fisher.information.normal}
\usage{
fisher.information.normal(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
fisher.information.normal
}
