% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.logistic.bootstrap.R
\name{gof.logistic.bootstrap}
\alias{gof.logistic.bootstrap}
\title{gof.logistic.bootstrap}
\usage{
gof.logistic.bootstrap(x, M = 10000)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
gof.logistic.bootstrap
}
