% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.gamma.regression.R
\name{score.gamma.regression}
\alias{score.gamma.regression}
\title{score.gamma.regression}
\usage{
score.gamma.regression(y, x, theta.hat)
}
\arguments{
\item{theta.hat}{}
}
\value{

}
\description{
score.gamma.regression
}
