% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.gamma.regression.R
\name{score.gamma.regression}
\alias{score.gamma.regression}
\title{Score Function for Gamma Regression Model}
\usage{
score.gamma.regression(y, x, theta.hat)
}
\arguments{
\item{y}{the univariate response}

\item{x}{the design matrix}

\item{theta.hat}{MLEs of parameters}
}
\value{
Row 1 of the output is the shape component of the score. Rows 2 to p+1
are the components of the score corresponding to the regression coefficients.
}
\description{
Compute a p+1 by n matrix containing the components of the score function
in a Gamma regression of y on the n x p covariate matrix x.
The ith column of the output corresponds to the ith data point.
}
\examples{
set.seed(5)
n = 200  #(sample size)
p = 3
beta = c(3,4,5)
alpha = 3
x = rnorm(n*(p-1))
x = c(rep(1,n),x)
x = matrix(x,n,p)
scale=exp(x \%*\% beta) / alpha
y =rgamma(n,shape=alpha,scale=scale)
theta.hat = estimate.gamma.regression(x,y)

}
