% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.normal.regression.R
\name{score.normal.regression}
\alias{score.normal.regression}
\title{score.normal.regression}
\usage{
score.normal.regression(y, x, theta.hat)
}
\arguments{
\item{theta.hat}{}
}
\value{

}
\description{
score.normal.regression
}
